symlinks.sh: avoid creating symlinks inside directories
This change makes symlinks.sh idempotent; with the full name of the
to-be-created symlink specified, a re-run would create a symlink to the
$EFM2Riot/dist/cpu/$X inside $RIOT/cpu/$X, which would either be a
symlink already and thus create $EFM2Riot/dist/cpu/$X/$X, or (in case
there is a directory around in the source tree because that CPU exists
upstream or the directory was not removed after a branch switch) inside
that directory.

The updated version now complains about existing directories (eg.

    ln: failed to create symbolic link '/home/chrysn/git/RIOT/boards/slwstk6220a/slwstk6220a': File exists
    ln: failed to create symbolic link '/home/chrysn/git/RIOT/cpu/ezr32wg/ezr32wg': File exists

) but does not create symlinks in the wrong place.

The -F option was changed to -f because -F makes no sense here ("allow
the superuser to attempt to hard link directories"), probably -f was
meant, and this makes the script overwrite existing symlinks.